refactor: reorganize directory structure to follow DDD best practices
- Move from models/ to value-objects/ for better DDD pattern clarity
- Create proper directory structure for all contexts (authentication, api, data, log)
- Add entities/, aggregates/, events/, repositories/ directories for future use
- Update all import paths to reflect new structure
- Add README.md explaining the DDD directory structure
- Follow medium-scale project structure from docs/015-ddd-directory-best-practices.md
