+++ apps/api/src/contexts/authentication/domain/value-objects/AuthenticatedUser.test.ts
