+++ apps/api/src/contexts/authentication/domain/value-objects/index.ts
