+++ apps/api/src/contexts/data/domain/index.ts
@@ -1,6 +1,6 @@
-export * from './value-objects';
-export * from './services';
 // export * from './entities';
+// export * from './value-objects';
 // export * from './aggregates';
+// export * from './services';
 // export * from './events';
 // export * from './repositories';
